fix: Fix cartPages function, Modulization
